+++ tests/test_app.py
@@ -7,8 +7,8 @@ def test_explain():
     with Client(app) as client:
         request_body = {
             'code': 'print("Hello, World!")',
-            'session': 'testemail: alex@darklight.ai',
-            'organization': 'darklight-ai',
+            'session': 'testemail: alex@polytest.ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }
 
@@ -24,8 +24,8 @@ def test_generate_outputlanguage():
         request_body = {
             'explanation': 'This is a simple program that prints "Hello, World!" to the console.',
             'originalCode': 'print("Hello, World!")',
-            'session': 'testemail: alex@darklight.ai',
-            'organization': 'darklight-ai',
+            'session': 'testemail: alex@polytest.ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5',
             'output_language': output_language
         }
@@ -41,7 +41,7 @@ def test_auth():
         request_body = {
             'session': 'testemail: foo@bar.com',
             'code': 'print("Hello, World!")',
-            'organization': 'darklight-ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }
 
@@ -54,7 +54,7 @@ def test_auth():
         request_body = {
             'session': 'testemail: alexgo@gmail.com',
             'code': 'print("Hello, World!")',
-            'organization': 'darklight-ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }
 
@@ -65,9 +65,9 @@ def test_auth():
 
         # now test with an email of jkthecjer@gmail.com
         request_body = {
-            'session': 'testemail: jkthecjer@gmail.com',
+            'session': 'testemail: jkthecjer@polytest.ai',
             'code': 'print("Hello, World!")',
-            'organization': 'darklight-ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }
 
@@ -76,11 +76,11 @@ def test_auth():
 
         assert response.payload['statusCode'] == 200
 
-        # now test with an email of alex@darklight.ai
+        # now test with an email of alex@polytest.ai
         request_body = {
-            'session': 'testemail: alex@darklight.ai',
+            'session': 'testemail: alex@polytest.ai',
             'code': 'print("Hello, World!")',
-            'organization': 'darklight-ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }
         response = client.lambda_.invoke(
@@ -96,8 +96,8 @@ def test_testgen():
             'language': 'python',
             'framework': 'pytest',
             'code': 'print("Hello, World!")',
-            'session': 'testemail: alex@darklight.ai',
-            'organization': 'darklight-ai',
+            'session': 'testemail: alex@polytest.ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }
         response = client.lambda_.invoke(
@@ -110,8 +110,8 @@ def test_analyze():
     with Client(app) as client:
         request_body = {
             'code': 'print("Hello, World!")',
-            'session': 'testemail: alex@darklight.ai',
-            'organization': 'darklight-ai',
+            'session': 'testemail: alex@polytest.ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }
 
@@ -125,8 +125,8 @@ def test_compliance():
     with Client(app) as client:
         request_body = {
             'code': 'print("Hello, World!")',
-            'session': 'testemail: alex@darklight.ai',
-            'organization': 'darklight-ai',
+            'session': 'testemail: alex@polytest.ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }
 
@@ -140,8 +140,8 @@ def test_codeguidelines():
     with Client(app) as client:
         request_body = {
             'code': 'print("Hello, World!")',
-            'session': 'testemail: alex@darklight.ai',
-            'organization': 'darklight-ai',
+            'session': 'testemail: alex@polytest.ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }
 
@@ -155,8 +155,8 @@ def test_blueprint():
     with Client(app) as client:
         request_body = {
             'code': 'print("Hello, World!")',
-            'session': 'testemail: alex@darklight.ai',
-            'organization': 'darklight-ai',
+            'session': 'testemail: alex@polytest.ai',
+            'organization': 'polytest.ai',
             'version': '0.9.5'
         }